requirements.txt
tinkoff-invest-api==1.8.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-learn==1.3.0
ta==0.10.2
yfinance==0.2.18
plotly==5.15.0
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.8.5
asyncio==3.4.3
schedule==1.2.0
backtrader==1.9.78.123
pytest==7.4.0
jupyter==1.0.0
xlsxwriter==3.1.2